Inserido metodo ContaTempo para Game Over. Alterado arquivo __init__ para classes_metodos e inserido no mesmo diretorio de sistema.
